requirements.txt
cairocffi==1.3.0
CairoSVG==2.5.2
cffi==1.15.0
colorama==0.4.4
cssselect2==0.5.0
defusedxml==0.7.1
Pillow==9.0.1
pycparser==2.21
tinycss2==1.1.1
tqdm==4.63.0
webencodings==0.5.1
